+++ crates/tests/src/lib.rs
@@ -39,6 +39,8 @@ mod test_object;
 #[cfg(test)]
 #[cfg(all(target_vendor = "apple", feature = "unstable-simd"))]
 mod test_simd_return;
+#[cfg(test)]
+mod writeback;
 
 // Run some `objc2` doctests that require extra crates.
 #[cfg(target_vendor = "apple")]